Simplifying the configuration of Storage. Use a simpler version of ISqlConnectionFactory that can be overriden by users in more complex cases
