+++ src/OhioBox.Storage.MySql/ISqlConnectionFactory.cs
@@ -7,10 +7,5 @@ namespace OhioBox.Storage.MySql
 	public interface ISqlConnectionFactory
 	{
 		IDbConnection GetConnection<T>();
-		ISqlConnectionProvider GetConnectionProvider<T>();
-		IEnumerable<ISqlConnectionProvider> Providers { get; }
-		IList<KeyValuePair<string, string>> GetTypeFullNameAndAonnectionString();
-
-		void Add(string connectionString, IEnumerable<Type> types);
 	}
 }
