--- src/OhioBox.Storage.MySql/SqlConnectionProviderExtension.cs
@@ -1,13 +1,0 @@
-﻿using System.Data;
-
-namespace OhioBox.Storage.MySql
-{
-	public static class SqlConnectionProviderExtension
-	{
-		public static IDbConnection GetOpenConnection(this ISqlConnectionProvider provider)
-		{
-			var connection = provider.ProvideConnection();
-			return connection;
-		}
-	}
-}
